SOC-1785 | truncate gets two separated parameters and goes to main/utils folder
