+++ front/scripts/main/helpers/truncate.js
@@ -1,9 +1,12 @@
-import truncate from '../../mercury/utils/truncate';
+import truncate from '../utils/truncate';
 
 /**
  * @param {Array} params
  * @returns {string}
  */
 export default Ember.Helper.helper((params) => {
-	return truncate(params);
+	const text = params[0],
+		maxLength = params[1];
+
+	return truncate(text, maxLength);
 });
